requirement.txt
fastapi
uvicorn
python-docx
PyMuPDF
sentence-transformers
faiss-cpu
pydantic
